requirements.txt
python-binance==1.0.17
python-dotenv==1.0.1
pyTelegramBotAPI==4.14.0
websocket-client==1.7.0
Flask==3.1.1




